# Inline Default impls and trivial instance/device wrapper methods