I did not inline `load`, any of the wrappers that invoke `read_into_*_vector` helpers, or (for now) extensions methods. In theory this should be a nice clean overhead reduction at runtime. Data on buildtime impact would be great!