@MarijnS95 2022-03-29
So it seems my random yet completely unrepresentative workstation machine is now used for benchmarking :joy: 

Before:

```console
$ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      5.841 s ±  0.031 s    [User: 6.671 s, System: 0.373 s]
  Range (min … max):    5.786 s …  5.898 s    10 runs
```

After:

```console
$ hyperfine -p 'cargo clean -p ash' 'cargo b -p ash'
Benchmark 1: cargo b -p ash
  Time (mean ± σ):      5.724 s ±  0.026 s    [User: 6.435 s, System: 0.364 s]
  Range (min … max):    5.691 s …  5.782 s    10 runs
```

`~120ms`.